refactor: move impl to RepeatStep object
